omsBackend/requirements.txt
amqp==2.4.2
billiard==3.5.0.5
celery==4.2.1
certifi==2018.11.29
chardet==3.0.4
colorama==0.4.1
Django==2.1.4
django-celery-results==1.0.4
django-cors-headers==2.4.0
django-extra-fields==1.0.0
django-filter==2.0.0
django-python3-ldap==0.11.2
django-rest-auth==0.9.3
django-uuslug==1.1.8
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
dry-rest-permissions==0.1.10
idna==2.7
json-tools==0.4.1
kombu==4.4.0
ldap3==2.5
lxml==4.2.3
Pillow==5.4.1
pyasn1==0.4.4
pygame==1.9.4
PyJWT==1.7.1
pytesseract==0.2.6
python-slugify==1.2.6
pytz==2018.7
qiniu==7.2.2
requests==2.20.1
six==1.12.0
tesseract==0.1.3
Unidecode==1.0.23
urllib3==1.24.1
vine==1.2.0
